refactored schema.ts to typeDefs.ts to improve consitency
